Converted Nav to TypeScript and fixed bug with RandomiseOrder
